refactor: moved entities and constants around
