+++ backend/src/modules/application/enums/application-type.enum.ts
@@ -1,4 +1,4 @@
-export enum ApplicationType {
+export enum ApplicationTypes {
   TYPE_1 = 'Website independent',
   TYPE_2 = 'Colecteaza date din ONGHub',
   TYPE_3 = 'Necesita login simplu',
